pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api==2.2.0",
    "fastapi>=0.115.14",
    "google-genai>=1.24.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "svix>=1.68.0",
    "uvicorn>=0.35.0",
]
